refactor(fieldset): rewrite to agreed format
